requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.32
aiosqlite==0.20.0
chromadb==0.4.24
langchain==0.2.12
openai==1.40.3
matplotlib==3.9.0
google-api-python-client==2.143.0
aiohttp==3.9.5
numpy==1.26.4
pydantic==2.8.2
tenacity==8.5.0
cachetools==5.4.0
python-dotenv==1.0.1